Add files via upload
In many water places around the world, water waste is a widespread issue. To prevent water contamination, fish kills, and the extinction of various aquatic species, clean water and sanitation are crucial issues that must be closely monitored. 
The wastes on the water areas are not only causing trouble on the people around it but also on the living organisms in the water. This project aims to create an automated waste detection system using image processing that will gather the waste images or a video from a camera and put the waste material to the bin. The purpose of this floating waste collector project is to clean the waste that gets accumulated on the surface of water bodies that results in keeping the water clean hence decreasing the pollution.